build: fix tree shaking for front matter related functions
