+++ src/language-markdown/parser-markdown.js
@@ -4,7 +4,7 @@ const remarkParse = require("remark-parse");
 const unified = require("unified");
 const remarkMath = require("remark-math");
 const footnotes = require("remark-footnotes");
-const { parse: parseFrontMatter } = require("../utils/front-matter");
+const parseFrontMatter = require("../utils/front-matter/parse");
 const pragma = require("./pragma");
 const { locStart, locEnd } = require("./loc");
 const { mapAst, INLINE_NODE_WRAPPER_TYPES } = require("./utils");
